+++ turbo/rpchelper/subscription.go
@@ -5,16 +5,14 @@ import (
 	"sync"
 )
 
+// a simple interface for subscriptions for rpc helper
 type Sub[T any] interface {
-	// Send should sends T to the subscriber. should be safe to send after Close()
 	Send(T)
-	// Close should mark the subscription closed. It should be safe to call Close multiple times
 	Close()
 }
 
 type chan_sub[T any] struct {
 	ch chan T
-	in chan T
 
 	closed chan struct{}
 
@@ -24,47 +22,26 @@ type chan_sub[T any] struct {
 	c sync.Cond
 }
 
-func NewChanSub[T any](ch chan T) *chan_sub[T] {
+// buffered channel
+func newChanSub[T any](size int) *chan_sub[T] {
+	// set min size to 8.
+	if size < 8 {
+		size = 8
+	}
 	o := &chan_sub[T]{}
-	o.ch = ch
-	o.in = make(chan T)
+	o.ch = make(chan T, size)
 	o.closed = make(chan struct{})
 	o.ctx, o.cn = context.WithCancel(context.Background())
-	go o.loop()
 	return o
 }
-
-func (s *chan_sub[T]) loop() {
-	for {
-		select {
-		// if canceled, exit loop
-		case <-s.ctx.Done():
-			return
-			// if there are any entries in the input buffer, grab
-		case v := <-s.in:
-			select {
-			// send to output channel
-			case s.ch <- v:
-			case <-s.ctx.Done():
-				return
-			}
-		}
-	}
-}
-
 func (s *chan_sub[T]) Send(x T) {
 	select {
 	// if the output buffer is empty, send
 	case s.ch <- x:
 		// if sub is canceled, dispose message
 	case <-s.ctx.Done():
+		// the sub is overloaded, dispose message
 	default:
-		select {
-		// if input buffer is empty, send
-		case s.in <- x:
-			// if sub is canceled, dispose message
-		case <-s.ctx.Done():
-		}
 	}
 }
 func (s *chan_sub[T]) Close() {
@@ -80,9 +57,76 @@ func (s *chan_sub[T]) Close() {
 	case s.closed <- struct{}{}:
 		// but it is not possible for multiple goroutines to get to this point
 		// drain the channel
-		for range s.ch {
-		}
-		close(s.ch)
+	//for range s.ch {
+	//}
+	//close(s.ch)
 	default:
 	}
+}
+
+func NewSyncMap[K comparable, T any]() *SyncMap[K, T] {
+	return &SyncMap[K, T]{
+		m: make(map[K]T),
+	}
+}
+
+type SyncMap[K comparable, T any] struct {
+	m map[K]T
+	sync.RWMutex
+}
+
+func (m *SyncMap[K, T]) Get(k K) (res T, ok bool) {
+	m.RLock()
+	defer m.RUnlock()
+	res, ok = m.m[k]
+	return res, ok
+}
+
+func (m *SyncMap[K, T]) Put(k K, v T) (T, bool) {
+	m.Lock()
+	defer m.Unlock()
+	_, ok := m.m[k]
+	m.m[k] = v
+	return m.m[k], ok
+}
+
+func (m *SyncMap[K, T]) Do(k K, fn func(T, bool) (T, bool)) (after T, ok bool) {
+	m.Lock()
+	defer m.Unlock()
+	val, ok := m.m[k]
+	nv, save := fn(val, ok)
+	if save {
+		m.m[k] = nv
+	}
+	return nv, ok
+}
+
+func (m *SyncMap[K, T]) DoAndStore(k K, fn func(t T, ok bool) T) (after T, ok bool) {
+	return m.Do(k, func(t T, b bool) (T, bool) {
+		res := fn(t, b)
+		return res, true
+	})
+}
+
+func (m *SyncMap[K, T]) Range(fn func(k K, v T) error) error {
+	m.RLock()
+	defer m.RUnlock()
+	for k, v := range m.m {
+		if err := fn(k, v); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (m *SyncMap[K, T]) Delete(k K) (T, bool) {
+	var t T
+	m.Lock()
+	defer m.Unlock()
+	val, ok := m.m[k]
+	if !ok {
+		return t, false
+	}
+	delete(m.m, k)
+	return val, true
 }
